requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
python-dotenv==1.0.1
pydantic==2.7.3
aiosqlite==0.20.0
httpx==0.27.0
numpy>=1.24.0
pytest-asyncio==0.23.7
# Geospatial
rasterio>=1.3.10
geopandas>=0.14.4
rasterstats>=0.19.0
fiona>=1.9.6
shapely>=2.0.4
pyproj>=3.6.1
python-multipart>=0.0.9
pyarrow>=15.0.0
anthropic>=0.39.0
